fix a possible bug in spinwave.m
When nSlice is larger than nHkl (e.g. for nMagExt > 1000), the memory management will assign nSlice = nHkl (line629). However, in this case, the floor command in line656 might be afflicted with even tiny numeric errors and leads to irregular hklIdx. For example, for nSlice = nHkl = 101, hklIdx will contain repeated 29 and bring up a error message in the calculation. Now a tiny numer of 1E-7 is added to the floor command. This can correct the slight numerical error and produce regular hklIdx. Please check whether this is the case.